Add DriverConnection class
