+++ server/drivers/driver-connection.js
@@ -1,0 +1,152 @@
+const uuid = require('uuid/v4');
+const drivers = require('./drivers');
+const renderConnection = require('./render-connection');
+const appLog = require('../lib/appLog');
+const getMeta = require('../lib/getMeta');
+
+/**
+ * Driver connection represents a driver and connection pairing, resulting in a connected database connection.
+ * In the future, it will optionally support a persisted connection,
+ * maintaining state about whether a connection is connected or not.
+ * These persisted connections will be reused across runQuery calls.
+ */
+class DriverConnection {
+  /**
+   * @param {object} connection
+   * @param {object} [user] - user to run query under. may not be provided if chart links turned on
+   */
+  constructor(connection, user) {
+    this.id = uuid();
+    this.connection = connection;
+    this.driver = drivers[connection.driver];
+    this.user = user;
+  }
+
+  /**
+   * Run query using driver implementation of connection
+   * @param {*} query
+   * @returns {Promise}
+   */
+  async runQuery(query) {
+    const connection = this.connection;
+    const driver = this.driver;
+    const user = this.user;
+
+    const finalResult = {
+      id: uuid.v4(),
+      cacheKey: null,
+      startTime: new Date(),
+      stopTime: null,
+      queryRunTime: null,
+      fields: [],
+      incomplete: false,
+      meta: {},
+      rows: []
+    };
+
+    const renderedConnection = renderConnection(connection, user);
+    const connectionName = renderedConnection.name;
+
+    appLog.debug(
+      {
+        originalConnection: connection,
+        renderedConnection,
+        user
+      },
+      'Rendered connection for user'
+    );
+
+    const queryContext = {
+      driver: connection.driver,
+      userId: user && user._id,
+      userEmail: user && user.email,
+      connectionId: connection._id,
+      connectionName,
+      query,
+      startTime: finalResult.startTime
+    };
+
+    appLog.info(queryContext, 'Running query');
+
+    let results;
+    try {
+      results = await driver.runQuery(query, renderedConnection);
+    } catch (error) {
+      // It is logged INFO because it isn't necessarily a server/application error
+      // It could just be a bad query
+      appLog.info(
+        { ...queryContext, error: error.toString() },
+        'Error running query'
+      );
+
+      // Rethrow the error
+      // The error here is something expected and should be shown to user
+      throw error;
+    }
+
+    let { rows, incomplete, suppressedResultSet } = results;
+
+    if (!Array.isArray(rows)) {
+      appLog.warn(
+        {
+          driver: connection.driver,
+          connectionId: connection._id,
+          connectionName,
+          query
+        },
+        'Expected rows to be an array but received %s.',
+        typeof rows
+      );
+      rows = [];
+    }
+
+    finalResult.incomplete = Boolean(incomplete);
+    finalResult.suppressedResultSet = Boolean(suppressedResultSet);
+    finalResult.rows = rows;
+    finalResult.stopTime = new Date();
+    finalResult.queryRunTime = finalResult.stopTime - finalResult.startTime;
+    finalResult.meta = getMeta(rows);
+    finalResult.fields = Object.keys(finalResult.meta);
+
+    appLog.info(
+      {
+        ...queryContext,
+        stopTime: finalResult.stopTime,
+        queryRunTime: finalResult.queryRunTime,
+        rowCount: rows.length,
+        incomplete: finalResult.incomplete,
+        suppressedResultSet: finalResult.suppressedResultSet
+      },
+      'Query finished'
+    );
+
+    return finalResult;
+  }
+
+  /**
+   * Test connection passed in using the driver implementation
+   * As long as promise resolves without error
+   * it is considered a successful connection config
+   */
+  testConnection() {
+    const renderedConnection = renderConnection(this.connection, this.user);
+    return this.driver.testConnection(renderedConnection);
+  }
+
+  /**
+   * Gets schema (sometimes called schemaInfo) for connection
+   * This data is used by client to build schema tree in editor sidebar
+   * @returns {Promise}
+   */
+  getSchema() {
+    // Increase the max rows without modifiying original connection
+    const connectionMaxed = {
+      ...this.connection,
+      maxRows: Number.MAX_SAFE_INTEGER
+    };
+    const renderedConnection = renderConnection(connectionMaxed, this.user);
+    return this.driver.getSchema(renderedConnection);
+  }
+}
+
+module.exports = DriverConnection;
