+++ server/drivers/index.js
@@ -1,9 +1,7 @@
-const uuid = require('uuid');
 const utils = require('./utils');
-const getMeta = require('../lib/getMeta');
-const appLog = require('../lib/appLog');
 const renderConnection = require('./render-connection');
 const drivers = require('./drivers');
+const DriverConnection = require('./driver-connection');
 
 /**
  * Run query using driver implementation of connection
@@ -13,97 +11,8 @@ const drivers = require('./drivers');
  * @returns {Promise}
  */
 async function runQuery(query, connection, user) {
-  const driver = drivers[connection.driver];
-
-  const finalResult = {
-    id: uuid.v4(),
-    cacheKey: null,
-    startTime: new Date(),
-    stopTime: null,
-    queryRunTime: null,
-    fields: [],
-    incomplete: false,
-    meta: {},
-    rows: []
-  };
-
-  const renderedConnection = renderConnection(connection, user);
-  const connectionName = renderedConnection.name;
-
-  appLog.debug(
-    {
-      originalConnection: connection,
-      renderedConnection,
-      user
-    },
-    'Rendered connection for user'
-  );
-
-  const queryContext = {
-    driver: connection.driver,
-    userId: user && user._id,
-    userEmail: user && user.email,
-    connectionId: connection._id,
-    connectionName,
-    query,
-    startTime: finalResult.startTime
-  };
-
-  appLog.info(queryContext, 'Running query');
-
-  let results;
-  try {
-    results = await driver.runQuery(query, renderedConnection);
-  } catch (error) {
-    // It is logged INFO because it isn't necessarily a server/application error
-    // It could just be a bad query
-    appLog.info(
-      { ...queryContext, error: error.toString() },
-      'Error running query'
-    );
-
-    // Rethrow the error
-    // The error here is something expected and should be shown to user
-    throw error;
-  }
-
-  let { rows, incomplete, suppressedResultSet } = results;
-
-  if (!Array.isArray(rows)) {
-    appLog.warn(
-      {
-        driver: connection.driver,
-        connectionId: connection._id,
-        connectionName,
-        query
-      },
-      'Expected rows to be an array but received %s.',
-      typeof rows
-    );
-    rows = [];
-  }
-
-  finalResult.incomplete = Boolean(incomplete);
-  finalResult.suppressedResultSet = Boolean(suppressedResultSet);
-  finalResult.rows = rows;
-  finalResult.stopTime = new Date();
-  finalResult.queryRunTime = finalResult.stopTime - finalResult.startTime;
-  finalResult.meta = getMeta(rows);
-  finalResult.fields = Object.keys(finalResult.meta);
-
-  appLog.info(
-    {
-      ...queryContext,
-      stopTime: finalResult.stopTime,
-      queryRunTime: finalResult.queryRunTime,
-      rowCount: rows.length,
-      incomplete: finalResult.incomplete,
-      suppressedResultSet: finalResult.suppressedResultSet
-    },
-    'Query finished'
-  );
-
-  return finalResult;
+  const driverConnection = new DriverConnection(connection, user);
+  return driverConnection.runQuery(query);
 }
 
 /**
@@ -114,9 +23,8 @@ async function runQuery(query, connection, user) {
  * @param {object} user
  */
 function testConnection(connection, user) {
-  const driver = drivers[connection.driver];
-  const renderedConnection = renderConnection(connection, user);
-  return driver.testConnection(renderedConnection);
+  const driverConnection = new DriverConnection(connection, user);
+  return driverConnection.testConnection();
 }
 
 /**
@@ -127,10 +35,8 @@ function testConnection(connection, user) {
  * @returns {Promise}
  */
 function getSchema(connection, user) {
-  connection.maxRows = Number.MAX_SAFE_INTEGER;
-  const driver = drivers[connection.driver];
-  const renderedConnection = renderConnection(connection, user);
-  return driver.getSchema(renderedConnection);
+  const driverConnection = new DriverConnection(connection, user);
+  return driverConnection.getSchema();
 }
 
 /**
